Guardar cambios antes de cambiar de rama
